+++ devicemanagement/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 ### Fixed
+- [CI] Adapt pipeline and conanfile.py to gitlab.com
 ### Security
 
 ## [0.11.0] - 2023-06-02